acp/requirements.txt
django>=5.1.7
djangorestframework>=3.14.0
drf-spectacular>=0.27.1
django-cors-headers>=4.3.1
python-dotenv>=1.0.1
gunicorn>=21.2.0
pyjwt[crypto]>=2.8.0
requests>=2.31.0
scrapybara>=0.1.0
anthropic>=0.18.1
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
drf-nested-routers>=0.93.4
cryptography>=42.0.0
pandas>=2.1.0
openpyxl>=3.1.2
crewai>=0.51.1
virtuals-acp>=0.1.16
langchain>=0.2.10
langchain-chroma>=0.1.2
chromadb>=0.5.3
langchain-groq>=0.1.4